% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_BHalpha.R
\name{edit_param_BHalpha_sp}
\alias{edit_param_BHalpha_sp}
\title{edit_param_mum_sp}
\usage{
edit_param_BHalpha_sp(bio.prm, factor, species)
}
\arguments{
\item{bio.prm}{--> biology prm files}

\item{factor}{--> factor to multiply the old value of BHalpha}

\item{species}{--> species code in the Atlantis configuration}
}
\value{
bio.lines with BHalpha of species multiplied by factor, vector with X elements, X = number of lines of biology prm files.
Each element is a line of biology prm files
}
\description{
edit_param_mum_sp
}
