% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_KDENR.R
\name{edit_param_KDENR_sp}
\alias{edit_param_KDENR_sp}
\title{edit_param_KDENR_sp}
\usage{
edit_param_KDENR_sp(bio.lines, factor, species)
}
\arguments{
\item{bio.lines}{}

\item{factor}{--> factor to multiply the old value of KDENR}

\item{species}{--> species code in the Atlantis configuration}
}
\value{
bio.lines, vector with X elements, X = number of lines of biology prm files.
Each element is a line of biology prm files
}
\description{
edit_param_KDENR_sp
}
