% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_mq.R
\name{edit_param_mq_sp}
\alias{edit_param_mq_sp}
\title{edit_param_mq_sp}
\usage{
edit_param_mq_sp(bio.lines, factor, species)
}
\arguments{
\item{factor}{--> factor to multiply the old value of BHalpha}

\item{species}{--> species code in the Atlantis configuration}

\item{bio.prm}{--> biology prm files}
}
\value{
bio.lines with BHalpha of species multiplied by factor, vector with X elements, X = number of lines of biology prm files.
Each element is a line of biology prm files
}
\description{
edit_param_mq_sp
}
