% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_mum.R
\name{edit_param_mum_group_age}
\alias{edit_param_mum_group_age}
\title{edit_param_mum_group_age}
\usage{
edit_param_mum_group_age(
  bio.prm,
  new.mum,
  overwrite = F,
  new.file.name,
  single.group = F,
  vector.size = 10
)
}
\arguments{
\item{bio.prm}{Biology prm file path}

\item{new.mum}{Table of new mum for the species with vector.size age classes}

\item{overwrite}{Boolean. T = create a new biology prm files, called new.file.name}

\item{new.file.name}{Path of the new file}

\item{single.group}{Boolean. T if vector = 1 (useless then...)}

\item{vector.size}{Size of the age class vector to edit}
}
\value{
Nothing, write a new biology prm files
}
\description{
edit_param_mum_group_age
}
