% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abund.R
\name{extract_abundance_grp}
\alias{extract_abundance_grp}
\title{extract_abundance_grp}
\usage{
extract_abundance_grp(path, prefix = NULL)
}
\arguments{
\item{prefix}{}
}
\description{
extract_abundance_grp
}
