% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abund.R
\name{extract_abundance_sp}
\alias{extract_abundance_sp}
\title{extract_abundance_sp}
\usage{
extract_abundance_sp(path, prefix = NULL, species)
}
\arguments{
\item{species}{}
}
\description{
extract_abundance_sp
}
