% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{extract_biomass_main}
\alias{extract_biomass_main}
\title{extract_biomass_main}
\usage{
extract_biomass_main(path, prefix = NULL, fg.file)
}
\arguments{
\item{sp}{}
}
\description{
extract_biomass_main
}
