% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{extract_catch_main}
\alias{extract_catch_main}
\title{extract_catch_main}
\usage{
extract_catch_main(path, prefix = NULL, fg.file, fishery = F)
}
\arguments{
\item{sp}{}
}
\description{
extract_catch_main
}
