% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Nbiomass.R
\name{extract_Nbiomass_main}
\alias{extract_Nbiomass_main}
\title{extract_Nbiomass_main}
\usage{
extract_Nbiomass_main(path, prefix = NULL, fg.file)
}
\arguments{
\item{fg.file}{}
}
\description{
extract_Nbiomass_main
}
