% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{open_main_nc}
\alias{open_main_nc}
\title{open_main_nc
Open the "XXX_OUT.nc" file}
\usage{
open_main_nc(path, prefix = NULL)
}
\arguments{
\item{prefix}{}
}
\description{
open_main_nc
Open the "XXX_OUT.nc" file
}
