% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_mq.R
\name{edit_param_mq_sp}
\alias{edit_param_mq_sp}
\title{edit_param_mq_sp}
\usage{
edit_param_mq_sp(bio.lines, factor, species)
}
\arguments{
\item{factor}{Factor to multiply the old value of mq}

\item{species}{Species code in the Atlantis configuration}

\item{bio.prm}{ReadLines(Biology prm files)}
}
\value{
bio.lines with mq of species multiplied by factor
}
\description{
edit_param_mq_sp
}
