% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_catch.R
\name{extract_catch_main}
\alias{extract_catch_main}
\title{extract_catch_main}
\usage{
extract_catch_main(path, prefix = NULL, fg.file, fishery = F, dt.timeserie)
}
\arguments{
\item{dt.timeserie}{}
}
\description{
extract_catch_main
}
