% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{read_atlantis}
\alias{read_atlantis}
\title{read_atlantis}
\usage{
read_atlantis(
  path,
  prefix = NULL,
  fg.file,
  fishery = F,
  spatial = F,
  N_only = F,
  ...
)
}
\arguments{
\item{path}{: the path of the atlantis folder}
}
\value{
: An object of class Atlantis
}
\description{
read_atlantis
}
